Demos/watchtower-demo/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
anthropic==0.40.0
requests==2.32.3
python-multipart==0.0.18
